[REF] l10n_br_fiscal,l10n_br_nfe: generic import wizard line
Extract the NFe-specific import wizard line model
(l10n_br_nfe.import_xml.products) into a generic
l10n_br_fiscal.document.import.wizard.line so it can be reused by other
fiscal document importers (CTe, etc.), as originally proposed in
KMEE's PR #3546 (but reimplemented cleanly on the current 16.0 wizard).

- New generic model holds the document-type agnostic data (XML product
  name/code, NCM/CFOP de-para, commercial UoM/qty/price, internal
  product/UoM/supplier matching) and the supplierinfo create/update
  logic, with _prepare_supplierinfo_vals() and _get_supplierinfo_price()
  hooks for specialization.
- imported_products_ids One2many moved up to the base fiscal import
  wizard so every document type inherits it.
- l10n_br_nfe extends the generic line in place, adding only the NFe
  specifics (commercial/tax unit split, ICMS/IPI) and injecting the
  partner UoM de-para into product.supplierinfo via the hook.
- Drop the dead "if False" UoM block and the stale PR #3546 TODO.
- Move the line model ACL from l10n_br_nfe to l10n_br_fiscal.
