+++ l10n_br_fiscal/wizards/__init__.py
@@ -1,3 +1,4 @@
 from . import base_wizard_mixin
 from . import document_status_wizard
+from . import document_import_wizard_line
 from . import document_import_wizard
