+++ l10n_br_nfe/security/ir.model.access.csv
@@ -1,5 +1,3 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-access_l10n_br_fiscal_import_products_wizard_user,l10n_br_nfe_import_xml_products_user,model_l10n_br_nfe_import_xml_products,l10n_br_nfe.group_user,1,0,0,0
-access_l10n_br_fiscal_import_products_wizard_manager,l10n_br_nfe_import_xml_products_manager,model_l10n_br_nfe_import_xml_products,l10n_br_nfe.group_manager,1,1,1,1
 access_l10n_br_nfe_mde_user,access_l10n_br_nfe_mde_user,model_l10n_br_nfe_mde,l10n_br_nfe.group_user,1,0,0,0
 access_l10n_br_nfe_mde_manager,access_l10n_br_nfe_mde_manager,model_l10n_br_nfe_mde,l10n_br_nfe.group_manager,1,1,1,1
