+++ l10n_br_nfe/wizards/document_import_wizard.py
@@ -19,12 +19,6 @@ from odoo.addons.l10n_br_fiscal.constants.fiscal import MODELO_FISCAL_NFE
 class DocumentImportWizard(models.TransientModel):
     _inherit = "l10n_br_fiscal.document.import.wizard"
 
-    imported_products_ids = fields.One2many(
-        string="Imported Products",
-        comodel_name="l10n_br_nfe.import_xml.products",
-        inverse_name="import_xml_id",
-    )
-
     nat_op = fields.Char(string="Natureza da Operação")
 
     @api.model
@@ -85,7 +79,7 @@ class DocumentImportWizard(models.TransientModel):
         product_ids = []
         for product in binding.infNFe.det:
             product_ids.append(
-                self.env["l10n_br_nfe.import_xml.products"]
+                self.env["l10n_br_fiscal.document.import.wizard.line"]
                 .create(self._prepare_imported_product_values(product))
                 .id
             )
@@ -96,29 +90,23 @@ class DocumentImportWizard(models.TransientModel):
         taxes = self._get_taxes_from_xml_product(product)
         supplier_id = self._search_product_supplier_by_product_code(product.prod.cProd)
         product_id = self._match_product(product.prod)
-        # if self.fiscal_operation_type == "in"
-        # and product_id and product_id.purchase_ok:
-        if False:  # seems former test is always true after you
-            # imported the product once -> it screws the UOM.
-            uom_id = product_id.uom_po_id
-        else:
+        uom_id = self.env["uom.uom"].search(
+            [
+                "|",
+                ("code", "=", product.prod.uCom),
+                ("code", "=", product.prod.uTrib),
+            ],
+            limit=1,
+        )
+        if not uom_id:  # search for alias
             uom_id = self.env["uom.uom"].search(
                 [
                     "|",
-                    ("code", "=", product.prod.uCom),
-                    ("code", "=", product.prod.uTrib),
+                    ("name", "=", product.prod.uCom),
+                    ("name", "=", product.prod.uTrib),
                 ],
                 limit=1,
             )
-            if not uom_id:  # search for alias
-                uom_id = self.env["uom.uom"].search(
-                    [
-                        "|",
-                        ("name", "=", product.prod.uCom),
-                        ("name", "=", product.prod.uTrib),
-                    ],
-                    limit=1,
-                )
 
         return {
             "product_name": product.prod.xProd,
@@ -331,64 +319,21 @@ class DocumentImportWizard(models.TransientModel):
         return parsed_xml
 
 
-# TODO transform that into a generic l10n_br_fiscal.document.line.import.wizard
-# as proposed https://github.com/OCA/l10n-brazil/pull/3546
-# https://github.com/kmee/l10n-brazil/blob/14.0-refactor-import-edoc
-# /l10n_br_nfe/wizards/document_line_import_wizard.py
-class NfeImportProducts(models.TransientModel):
-    _name = "l10n_br_nfe.import_xml.products"
-    _description = "Import XML NFe Products"
+class DocumentImportWizardLine(models.TransientModel):
+    """NFe specialization of the generic fiscal import wizard line.
 
-    product_name = fields.Char()
+    It only adds the NFe specific data: the commercial/tax unit split and
+    the ICMS/IPI taxes read from the XML. It also injects the partner UoM
+    de-para into the generated ``product.supplierinfo``.
+    """
 
-    uom_com = fields.Char(string="UOM Comercial")
-
-    quantity_com = fields.Float(string="Comercial Quantity")
-
-    price_unit_com = fields.Float(string="Comercial Price Unit")
+    _inherit = "l10n_br_fiscal.document.import.wizard.line"
 
     uom_trib = fields.Char(string="UOM Fiscal")
 
     quantity_trib = fields.Float()
 
     price_unit_trib = fields.Float(string="Fiscal Price Unit")
-
-    total = fields.Float()
-
-    import_xml_id = fields.Many2one(
-        comodel_name="l10n_br_fiscal.document.import.wizard"
-    )
-
-    product_code = fields.Char(string="XML Product Code")
-
-    product_id = fields.Many2one(
-        comodel_name="product.product",
-        string="Product Internal Reference",
-    )
-
-    product_supplier_id = fields.Many2one(
-        comodel_name="product.supplierinfo",
-        string="Product Supplier",
-    )
-
-    uom_internal = fields.Many2one(
-        comodel_name="uom.uom",
-        help="Internal UoM, equivalent to the comercial one in the document",
-    )
-
-    ncm_xml = fields.Char(string="XML NCM Code")
-
-    ncm_internal = fields.Char(
-        related="product_id.ncm_id.code",
-        string="Internal NCM Code",
-    )
-
-    cfop_xml = fields.Char(string="XML CFOP")
-
-    new_cfop_id = fields.Many2one(
-        comodel_name="l10n_br_fiscal.cfop",
-        string="Change CFOP",
-    )
 
     icms_percent = fields.Char(string="Alíquota ICMS")
 
@@ -398,53 +343,12 @@ class NfeImportProducts(models.TransientModel):
 
     ipi_value = fields.Char(string="IPI Value")
 
-    imported_partner_id = fields.Many2one(related="import_xml_id.partner_id")
-
-    uom_conversion_factor = fields.Float(string="UOM Conversion Factor", default=1)
-
-    def _find_or_create_product_supplierinfo(self):
-        for product in self:
-            if not product.product_id:
-                continue
-
-            if not product.product_supplier_id:
-                product._create_product_supplier()
-            else:
-                product._update_product_supplier()
-
-    def _create_product_supplier(self):
-        if self.uom_internal:
-            price = self.uom_internal._compute_price(
-                self.price_unit_com, self.product_id.uom_id
-            )
-        else:
-            price = self.product_id.lst_price
-
-        self.product_supplier_id = self.env["product.supplierinfo"].create(
+    def _prepare_supplierinfo_vals(self):
+        vals = super()._prepare_supplierinfo_vals()
+        vals.update(
             {
-                "product_id": self.product_id.id,
-                "product_name": self.product_name,
-                "product_code": self.product_code,
-                "price": price,
-                "partner_id": self.imported_partner_id.id,
                 "partner_uom_id": self.uom_internal.id,
                 "partner_uom_factor": self.uom_conversion_factor,
             }
         )
-        self.product_id.write(
-            {"seller_ids": [Command.link(self.product_supplier_id.id)]}
-        )
-
-    def _update_product_supplier(self):
-        self.product_supplier_id.write(
-            {
-                "product_id": self.product_id.id,
-                "product_name": self.product_name,
-                "product_code": self.product_code,
-                "price": self.uom_internal._compute_price(
-                    self.price_unit_com, self.product_id.uom_id
-                ),
-                "partner_uom_id": self.uom_internal.id,
-                "partner_uom_factor": self.uom_conversion_factor,
-            }
-        )
+        return vals
