+++ l10n_br_nfe/wizards/document_import_wizard.xml
@@ -29,8 +29,8 @@
     </record>
 
     <record id="l10n_br_nfe_import_xml_products_tree" model="ir.ui.view">
-        <field name="name">l10n_br_nfe.import_xml.products.tree</field>
-        <field name="model">l10n_br_nfe.import_xml.products</field>
+        <field name="name">l10n_br_fiscal.document.import.wizard.line.tree</field>
+        <field name="model">l10n_br_fiscal.document.import.wizard.line</field>
         <field name="arch" type="xml">
             <tree editable="top">
                 <field
